+++ tsconfig.json
@@ -1,9 +1,14 @@
 {
     "compilerOptions": {
         "strict": true,
+        "declaration": true,
         "target": "es6",
         "module": "commonjs",
         "moduleResolution": "node",
-        "outDir": "dist"
-    }
+        "outDir": "dist",
+        "rootDir": "src",
+    },
+    "include": [
+        "./src/**/*"
+    ]
 }
