+++ tsconfig.test.json
@@ -1,0 +1,10 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+        "rootDir": ".",
+    },
+    "include": [
+        "./src/**/*",
+        "./tests/**/*"
+    ]
+}
